Move validate bin orders to utils
